+++ lib/whisperTab/whisper.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:blurting/messageClass.dart';
+import 'package:blurting/class/messageClass.dart';
 import 'package:web_socket_channel/io.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
